requirements.txt
# Core dependencies
gradio>=4.0.0
google-genai>=0.3.0

# For logging and data processing
typing-extensions>=4.0.0

# Optional: For development
python-dotenv>=1.0.0
